requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
Pillow>=9.0.0
scikit-learn>=1.2.0
